refactor: make item quantities i32 with explicit validation (#26)
Follow-up to #24 per Ben's ruling: `amount`, `weight_grams`, `price_eb`
(and the derived `calculate_total_weight`, `carry_capacity`, `deadlift`)
become `i32` too — amounts get subtracted (spent ammo), and `amount *
weight_grams` products would otherwise need casts everywhere. This
removes the last usize↔i32 cast in the codebase.

The non-negativity guarantee that `usize` gave implicitly is now
**explicit validation with meaningful errors**:
- `Item::new` panics with e.g. `Item 'Broomstick': weight_grams must not
be negative, got -1500`
- Loading a TOML file with negative quantities fails cleanly via
`#[serde(try_from = "UncheckedItem")]` instead of constructing invalid
state

Two new tests cover both paths (26 total, all passing).

Stacked on #25 → merge order: #24, #25, then this one (GitHub retargets
automatically).

+++ src/armor.rs
@@ -111,9 +111,9 @@ impl fmt::Display for Armor {
 impl Armor {
     pub fn new(
         name: String,
-        amount: usize,
-        weight_grams: usize,
-        price_eb: usize,
+        amount: i32,
+        weight_grams: i32,
+        price_eb: i32,
         comment: String,
         protection_max: i32,
         protected_zones: Vec<HitZone>,
